Move mediaart crawler to union
